convert delayed siacoin outputs management
The new style switches to the same sort of 'it either exists or it
doesn't' model that's followed by all of the other outputs. When adding
or removing a delayed output, you call commitDelayedOutput. There is a
much more strictly enforced structure that the delayed output maps must
follow, and a lot more build.DEBUG checking to provide assurance that
everything is running as intended.

Because there was a bit of an overhaul, I'm really not sure whether this
cleaned up any actual bugs or not. But I am sure that the program is
much more sensitive/self-aware of its delayed outputs structure, and
much better suited to warn the programmers that something unusual has
happened.
