+++ modules/consensus.go
@@ -53,6 +53,8 @@ type ConsensusChange struct {
 	// diffs is 'DiffApply'.
 	SiafundOutputDiffs []SiafundOutputDiff
 
+	// DelayedSiacoinOutputDiffs contains the set of delayed siacoin output diffs that were applied to the consensus set in the recent change.
+
 	// SiafundPoolDiff is the siafund pool diff that was applied to the
 	// consensus set in the recent change.
 	SiafundPoolDiff SiafundPoolDiff
@@ -84,6 +86,13 @@ type SiafundOutputDiff struct {
 	Direction     DiffDirection
 	ID            types.SiafundOutputID
 	SiafundOutput types.SiafundOutput
+}
+
+type DelayedSiacoinOutputDiff struct {
+	Direction      DiffDirection
+	ID             types.SiacoinOutputID
+	SiacoinOutput  types.SiacoinOutput
+	MaturityHeight types.BlockHeight
 }
 
 // A SiafundPoolDiff contains the value of the siafundPool before the block
